+++ index.html
@@ -5,7 +5,7 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Document</title>
-    <link rel="stylesheet" href="./src/style/town.css" />
+    <link rel="stylesheet" href="./src/stylesheet/style.css" />
   </head>
   <body>
     <div class="town">
